% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doh.R
\docType{package}
\name{doh-package}
\alias{doh}
\alias{doh-package}
\title{doh: Easily print the last value, for those D'oh moments.}
\description{
Easily print the last value, for those D'oh moments, for example after assiging a long command to a variable, without printing it.
}
\details{
a helper function by Michael Pinus (michael.pinus AT gmail.com) written in blood and D'oh moments, over the years.
}
\keyword{internal}
