% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLast.R
\name{printLast}
\alias{printLast}
\title{A wrapper of print(.Last.value), to be used with a shortcut}
\usage{
printLast()
}
\value{
None. a print method is used instead
}
\description{
A wrapper of print(.Last.value), to be used with a shortcut
}
